Extract tileset_cache::loader to src/tileset_loader.{h,cpp}
- new src/tileset_loader.{h,cpp} owns loader class and helpers
- cata_tiles.cpp keeps load_tileset dispatcher and includes the new header
- cata_tiles.h forward-declares loader and drops the inline class body
- overlay-ordering global stays in cata_tiles.cpp scope
